+++ libs/libc/misc/lib_mutex.c
@@ -82,18 +82,18 @@ static bool nxmutex_is_reset(FAR mutex_t *mutex)
 
 int nxmutex_init(FAR mutex_t *mutex)
 {
-  int ret = _SEM_INIT(&mutex->sem, 0, 1);
+  int ret = nxsem_init(&mutex->sem, 0, 1);
 
   if (ret < 0)
     {
-      return _SEM_ERRVAL(ret);
+      return ret;
     }
 
   mutex->holder = NXMUTEX_NO_HOLDER;
 #ifdef CONFIG_PRIORITY_INHERITANCE
-  _SEM_SETPROTOCOL(&mutex->sem, SEM_TYPE_MUTEX | SEM_PRIO_INHERIT);
+  nxsem_set_protocol(&mutex->sem, SEM_TYPE_MUTEX | SEM_PRIO_INHERIT);
 #else
-  _SEM_SETPROTOCOL(&mutex->sem, SEM_TYPE_MUTEX);
+  nxsem_set_protocol(&mutex->sem, SEM_TYPE_MUTEX);
 #endif
   return ret;
 }
@@ -119,11 +119,11 @@ int nxmutex_init(FAR mutex_t *mutex)
 
 int nxmutex_destroy(FAR mutex_t *mutex)
 {
-  int ret = _SEM_DESTROY(&mutex->sem);
+  int ret = nxsem_destroy(&mutex->sem);
 
   if (ret < 0)
     {
-      return _SEM_ERRVAL(ret);
+      return ret;
     }
 
   mutex->holder = NXMUTEX_NO_HOLDER;
@@ -167,7 +167,7 @@ bool nxmutex_is_locked(FAR mutex_t *mutex)
   int cnt;
   int ret;
 
-  ret = _SEM_GETVALUE(&mutex->sem, &cnt);
+  ret = nxsem_get_value(&mutex->sem, &cnt);
 
   return ret >= 0 && cnt < 1;
 }
@@ -242,10 +242,10 @@ int nxmutex_trylock(FAR mutex_t *mutex)
   int ret;
 
   DEBUGASSERT(!nxmutex_is_hold(mutex));
-  ret = _SEM_TRYWAIT(&mutex->sem);
+  ret = nxsem_trywait(&mutex->sem);
   if (ret < 0)
     {
-      return _SEM_ERRVAL(ret);
+      return ret;
     }
 
   mutex->holder = _SCHED_GETTID();
@@ -291,11 +291,7 @@ int nxmutex_timedlock(FAR mutex_t *mutex, unsigned int timeout)
 
   do
     {
-      ret = _SEM_CLOCKWAIT(&mutex->sem, CLOCK_MONOTONIC, &rqtp);
-      if (ret < 0)
-        {
-          ret = _SEM_ERRVAL(ret);
-        }
+      ret = nxsem_clockwait(&mutex->sem, CLOCK_MONOTONIC, &rqtp);
     }
   while (ret == -EINTR || ret == -ECANCELED);
 
@@ -340,11 +336,10 @@ int nxmutex_unlock(FAR mutex_t *mutex)
 
   mutex->holder = NXMUTEX_NO_HOLDER;
 
-  ret = _SEM_POST(&mutex->sem);
+  ret = nxsem_post(&mutex->sem);
   if (ret < 0)
     {
       mutex->holder = _SCHED_GETTID();
-      ret = _SEM_ERRVAL(ret);
     }
 
   return ret;
