+++ libs/libc/semaphore/Make.defs
@@ -20,11 +20,9 @@
 
 # Add the semaphore C files to the build
 
-CSRCS += sem_init.c sem_getprotocol.c sem_getvalue.c
-
-ifneq ($(CONFIG_PRIORITY_INHERITANCE),y)
-CSRCS += sem_setprotocol.c
-endif
+CSRCS += sem_init.c sem_setprotocol.c sem_getprotocol.c sem_getvalue.c
+CSRCS += sem_destroy.c sem_wait.c sem_trywait.c sem_timedwait.c
+CSRCS += sem_clockwait.c sem_post.c
 
 # Add the semaphore directory to the build
