+++ libs/libc/semaphore/sem_getvalue.c
@@ -24,7 +24,6 @@
 
 #include <nuttx/config.h>
 
-#include <sys/types.h>
 #include <errno.h>
 
 #include <nuttx/semaphore.h>
