+++ libs/libc/semaphore/sem_setprotocol.c
@@ -94,6 +94,8 @@ int nxsem_set_protocol(FAR sem_t *sem, int protocol)
   return OK;
 }
 
+#endif /* !CONFIG_PRIORITY_INHERITANCE */
+
 /****************************************************************************
  * Name: sem_setprotocol
  *
@@ -144,5 +146,3 @@ int sem_setprotocol(FAR sem_t *sem, int protocol)
 
   return ret;
 }
-
-#endif /* !CONFIG_PRIORITY_INHERITANCE */
