+++ libs/libc/semaphore/sem_trywait.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * sched/semaphore/sem_destroy.c
+ * libs/libc/semaphore/sem_trywait.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -26,57 +26,51 @@
 
 #include <errno.h>
 
-#include "semaphore/semaphore.h"
+#include <nuttx/semaphore.h>
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: nxsem_destroy
+ * Name: sem_trywait
  *
  * Description:
- *   This function is used to destroy the un-named semaphore indicated by
- *   'sem'.  Only a semaphore that was created using nxsem_init() may be
- *   destroyed using nxsem_destroy(); the effect of calling nxsem_destroy()
- *   with a named semaphore is undefined.  The effect of subsequent use of
- *   the semaphore sem is undefined until sem is re-initialized by another
- *   call to nxsem_init().
- *
- *   The effect of destroying a semaphore upon which other processes are
- *   currently blocked is undefined.
+ *   This function locks the specified semaphore only if the semaphore is
+ *   currently not locked.  In either case, the call returns without
+ *   blocking.
  *
  * Input Parameters:
- *   sem - Semaphore to be destroyed.
+ *   sem - the semaphore descriptor
  *
  * Returned Value:
- *   This is an internal OS interface and should not be used by applications.
- *   It follows the NuttX internal error return policy:  Zero (OK) is
- *   returned on success.  A negated errno value is returned on failure.
+ *   Zero (OK) on success or -1 (ERROR) if unsuccessful. If this function
+ *   returns -1(ERROR), then the cause of the failure will be reported in
+ *   errno variable as:
+ *
+ *     EINVAL - Invalid attempt to get the semaphore
+ *     EAGAIN - The semaphore is not available.
  *
  ****************************************************************************/
 
-int nxsem_destroy(FAR sem_t *sem)
+int sem_trywait(FAR sem_t *sem)
 {
-  DEBUGASSERT(sem != NULL);
+  int ret;
 
-  /* There is really no particular action that we need
-   * take to destroy a semaphore.  We will just reset
-   * the count to some reasonable value (0) and release
-   * ownership.
-   *
-   * Check if other threads are waiting on the semaphore.
-   * In this case, the behavior is undefined.  We will:
-   * leave the count unchanged but still return OK.
-   */
-
-  if (sem->semcount >= 0)
+  if (sem == NULL)
     {
-      sem->semcount = 1;
+      set_errno(EINVAL);
+      return ERROR;
     }
 
-  /* Release holders of the semaphore */
+  /* Let nxsem_trywait do the real work */
 
-  nxsem_destroyholder(sem);
-  return OK;
+  ret = nxsem_trywait(sem);
+  if (ret < 0)
+    {
+      set_errno(-ret);
+      ret = ERROR;
+    }
+
+  return ret;
 }
