+++ libs/libc/semaphore/sem_wait.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * sched/semaphore/sem_destroy.c
+ * libs/libc/semaphore/sem_wait.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -26,57 +26,73 @@
 
 #include <errno.h>
 
-#include "semaphore/semaphore.h"
+#include <nuttx/cancelpt.h>
+#include <nuttx/semaphore.h>
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: nxsem_destroy
+ * Name: sem_wait
  *
  * Description:
- *   This function is used to destroy the un-named semaphore indicated by
- *   'sem'.  Only a semaphore that was created using nxsem_init() may be
- *   destroyed using nxsem_destroy(); the effect of calling nxsem_destroy()
- *   with a named semaphore is undefined.  The effect of subsequent use of
- *   the semaphore sem is undefined until sem is re-initialized by another
- *   call to nxsem_init().
- *
- *   The effect of destroying a semaphore upon which other processes are
- *   currently blocked is undefined.
+ *   This function attempts to lock the semaphore referenced by 'sem'.  If
+ *   the semaphore value is (<=) zero, then the calling task will not return
+ *   until it successfully acquires the lock.
  *
  * Input Parameters:
- *   sem - Semaphore to be destroyed.
+ *   sem - Semaphore descriptor.
  *
  * Returned Value:
- *   This is an internal OS interface and should not be used by applications.
- *   It follows the NuttX internal error return policy:  Zero (OK) is
- *   returned on success.  A negated errno value is returned on failure.
+ *   This function is a standard, POSIX application interface.  It returns
+ *   zero (OK) if successful.  Otherwise, -1 (ERROR) is returned and
+ *   the errno value is set appropriately.  Possible errno values include:
+ *
+ *   - EINVAL:  Invalid attempt to get the semaphore
+ *   - EINTR:   The wait was interrupted by the receipt of a signal.
  *
  ****************************************************************************/
 
-int nxsem_destroy(FAR sem_t *sem)
+int sem_wait(FAR sem_t *sem)
 {
-  DEBUGASSERT(sem != NULL);
+  int errcode;
+  int ret;
 
-  /* There is really no particular action that we need
-   * take to destroy a semaphore.  We will just reset
-   * the count to some reasonable value (0) and release
-   * ownership.
-   *
-   * Check if other threads are waiting on the semaphore.
-   * In this case, the behavior is undefined.  We will:
-   * leave the count unchanged but still return OK.
-   */
-
-  if (sem->semcount >= 0)
+  if (sem == NULL)
     {
-      sem->semcount = 1;
+      set_errno(EINVAL);
+      return ERROR;
     }
 
-  /* Release holders of the semaphore */
+  /* sem_wait() is a cancellation point */
 
-  nxsem_destroyholder(sem);
+  if (enter_cancellation_point())
+    {
+#ifdef CONFIG_CANCELLATION_POINTS
+      /* If there is a pending cancellation, then do not perform
+       * the wait.  Exit now with ECANCELED.
+       */
+
+      errcode = ECANCELED;
+      goto errout_with_cancelpt;
+#endif
+    }
+
+  /* Let nxsem_wait() do the real work */
+
+  ret = nxsem_wait(sem);
+  if (ret < 0)
+    {
+      errcode = -ret;
+      goto errout_with_cancelpt;
+    }
+
+  leave_cancellation_point();
   return OK;
+
+errout_with_cancelpt:
+  set_errno(errcode);
+  leave_cancellation_point();
+  return ERROR;
 }
