+++ sched/semaphore/sem_clockwait.c
@@ -34,7 +34,6 @@
 #include <nuttx/irq.h>
 #include <nuttx/arch.h>
 #include <nuttx/wdog.h>
-#include <nuttx/cancelpt.h>
 
 #include "sched/sched.h"
 #include "clock/clock.h"
@@ -215,74 +214,5 @@ int nxsem_clockwait_uninterruptible(FAR sem_t *sem, clockid_t clockid,
     }
   while (ret == -EINTR);
 
-  return ret;
-}
-
-/****************************************************************************
- * Name: sem_clockwait
- *
- * Description:
- *   This function will lock the semaphore referenced by sem as in the
- *   sem_wait() function. However, if the semaphore cannot be locked without
- *   waiting for another process or thread to unlock the semaphore by
- *   performing a sem_post() function, this wait will be terminated when the
- *   specified timeout expires.
- *
- *   The timeout will expire when the absolute time specified by abstime
- *   passes, as measured by the clock on which timeouts are based (that is,
- *   when the value of that clock equals or exceeds abstime), or if the
- *   absolute time specified by abstime has already been passed at the
- *   time of the call.
- *
- * Input Parameters:
- *   sem     - Semaphore object
- *   clockid - The timing source to use in the conversion
- *   abstime - The absolute time to wait until a timeout is declared.
- *
- * Returned Value:
- *   Zero (OK) is returned on success.  On failure, -1 (ERROR) is returned
- *   and the errno is set appropriately:
- *
- *   EINVAL    The sem argument does not refer to a valid semaphore.  Or the
- *             thread would have blocked, and the abstime parameter specified
- *             a nanoseconds field value less than zero or greater than or
- *             equal to 1000 million.
- *   ETIMEDOUT The semaphore could not be locked before the specified timeout
- *             expired.
- *   EDEADLK   A deadlock condition was detected.
- *   EINTR     A signal interrupted this function.
- *   ECANCELED May be returned if the thread is canceled while waiting.
- *
- ****************************************************************************/
-
-int sem_clockwait(FAR sem_t *sem, clockid_t clockid,
-                  FAR const struct timespec *abstime)
-{
-  int ret;
-
-  /* Verify the input parameters and, in case of an error, set
-   * errno appropriately.
-   */
-
-  if (sem == NULL || abstime == NULL)
-    {
-      set_errno(EINVAL);
-      return ERROR;
-    }
-
-  /* sem_timedwait() is a cancellation point */
-
-  enter_cancellation_point();
-
-  /* Let nxsem_timedout() do the work */
-
-  ret = nxsem_clockwait(sem, clockid, abstime);
-  if (ret < 0)
-    {
-      set_errno(-ret);
-      ret = ERROR;
-    }
-
-  leave_cancellation_point();
   return ret;
 }
