+++ sched/semaphore/sem_post.c
@@ -205,54 +205,3 @@ int nxsem_post(FAR sem_t *sem)
 
   return OK;
 }
-
-/****************************************************************************
- * Name: sem_post
- *
- * Description:
- *   When a task has finished with a semaphore, it will call sem_post().
- *   This function unlocks the semaphore referenced by sem by performing the
- *   semaphore unlock operation on that semaphore.
- *
- *   If the semaphore value resulting from this operation is positive, then
- *   no tasks were blocked waiting for the semaphore to become unlocked; the
- *   semaphore is simply incremented.
- *
- *   If the value of the semaphore resulting from this operation is zero,
- *   then one of the tasks blocked waiting for the semaphore shall be
- *   allowed to return successfully from its call to nxsem_wait().
- *
- * Input Parameters:
- *   sem - Semaphore descriptor
- *
- * Returned Value:
- *   This function is a standard, POSIX application interface.  It will
- *   return zero (OK) if successful.  Otherwise, -1 (ERROR) is returned and
- *   the errno value is set appropriately.
- *
- * Assumptions:
- *   This function may be called from an interrupt handler.
- *
- ****************************************************************************/
-
-int sem_post(FAR sem_t *sem)
-{
-  int ret;
-
-  /* Make sure we were supplied with a valid semaphore. */
-
-  if (sem == NULL)
-    {
-      set_errno(EINVAL);
-      return ERROR;
-    }
-
-  ret = nxsem_post(sem);
-  if (ret < 0)
-    {
-      set_errno(-ret);
-      ret = ERROR;
-    }
-
-  return ret;
-}
