+++ sched/semaphore/sem_trywait.c
@@ -105,46 +105,3 @@ int nxsem_trywait(FAR sem_t *sem)
   leave_critical_section(flags);
   return ret;
 }
-
-/****************************************************************************
- * Name: sem_trywait
- *
- * Description:
- *   This function locks the specified semaphore only if the semaphore is
- *   currently not locked.  In either case, the call returns without
- *   blocking.
- *
- * Input Parameters:
- *   sem - the semaphore descriptor
- *
- * Returned Value:
- *   Zero (OK) on success or -1 (ERROR) if unsuccessful. If this function
- *   returns -1(ERROR), then the cause of the failure will be reported in
- *   errno variable as:
- *
- *     EINVAL - Invalid attempt to get the semaphore
- *     EAGAIN - The semaphore is not available.
- *
- ****************************************************************************/
-
-int sem_trywait(FAR sem_t *sem)
-{
-  int ret;
-
-  if (sem == NULL)
-    {
-      set_errno(EINVAL);
-      return ERROR;
-    }
-
-  /* Let nxsem_trywait do the real work */
-
-  ret = nxsem_trywait(sem);
-  if (ret < 0)
-    {
-      set_errno(-ret);
-      ret = ERROR;
-    }
-
-  return ret;
-}
